requirement.txt
ebooklib
download
pillow
uuid
bs4
